Behave differently for non-ARA regions
Previously, we would always run the NLCD+ARA request,
which ended up failing when outside the ARA region.
Now we check if the AoI is wholly contained within the
ARA Perimeter and only run NLCD+ARA in that case.
Otherwise we run only NLCD.

The analyze_nlcd task is similarly updated to behave
differently when receiving a tuple of values (indicating
the presence of ARA results) vs not, sending back nulls
instead.

We use null vs 0 here to differentiate absence of ARA
values within the raster with querying outside the bounds.