Add ARA output to Analyze Land tab
The output, shown in square km, will be shown in the
Analyze Land tab if the area of interest is within
the ARA Perimeter.

In the code, we inspect all the values for ARA for
all the land types. If some of them are NULL we treat
it as 0, and still sum up all the values to a total.
The value will be shown even if it is 0.

Only when all the values are NULL that we take it as
a signal that the AoI must be outside the ARA Perimeter,
in which case the entire line is hidden.