Switch Climate Analysis to use Multi-operation
To reduce Celery overhead and the possibility of unchecked
`chord_unlock` calls, we switch Climate analysis to the
multioperation geoprocessing call.

`analyze_climate` and `collect_climate` tasks are merged
into one, to expect the new input. Their collective output
remains the same as before.

The view is considerably simplified. Geoprocessing config
is updated to remove the solo calls to `ppt` and `tmean`,
and to have a single `climate` call that gathers data from
all 24 rasters.