Refactor analyze models to allow for multiple analysis per tab.
The motivation for this refactor is so that the Protected Lands
analysis, which is very similar to the NLCD analysis, can be on the
same "Land" tab in the Analysis results view of the app. Previously,
analysis tabs were tied to on geoprocessing API call and result
set, through the AnalyzeTaskModel. This refactor introduces a level of
hierarchy in-between, the AnalyzeTaskGroupModel, which allows tabs to
represent a group of tasks instead of strictly a single task.

The protected lands task is introduced here. The UI changes that take
advantage of this refactor to add Protected Lands to the Land tab is
contained in a subsequent commit.

Refs #3203