Create a selector to switch between analysis task in analysis tabs.
This refactor introduces a new selector that can live at the top of an
AnalyzeResultView which will switch the active task of the
AnalyzeTaskGroupModel. As there are now multiple selectors, the
selector that was controlling teh variable choice for the Climate tab
was renamed to VarSelectorView. AnalyzeResultViews are initialized
with their taskGroup in order to have access to changing the active
task on the model.

Refs #3203