Add Total row to Land analysis table, others
Previously the Active River Area total was displayed as
text below the table. Now that it is split across NLCD
values, it is more sensible to display its total as an
aggregate row at the bottom of the table.

Since the Land table is reused for Protected Lands and
Soil, a total row is added to those tables as well. This
total is included in the CSV export.

The NLCD Code column is excluded from totaling.

The Coverage totals to 100% for Land and Soil, but to
less than 100% for Protected Lands in most cases. This
is a useful number to expose, as it is useful to know
the total percent area of an area of interest that is
under protection.

If the area of interest is outside the Mid-Atlantic
North-East region, the Active River Area total will
read "No Data".