Better handling of ARA NODATA
Previously we would use a roughly drawn outline of the ARA
raster to see if the AoI fit within it, and only analyze
ARA if so. For shapes that were on the edge and did not fit
entirely within the outline, the front-end would show No Data
values, even though turning the visual layer on would show
data. This was confusing and incorrect.

The outline was added because requesting ARA data outside the
bounds would throw an exception in the geoprocessing service.
This has been amended in the 4.0.2 release, which now sends
back an empty tile instead of throwing an exception.

For edge cases, where some of the AoI has real ARA values and
the rest does not, we treat the NODATA values as 0.

For cases completely outside ARA, where there is not one real
value, we treat NODATA as null, which shows up as No Data in
the front-end.