Add weather stations to Observation tab.
This commit adds weather stations to the observation tab, which allows
the user to show all weather stations on the map. There is a new
endpoint that returns each weather station and properties stored in
the ms_weather_station table. The markers are based on the CUASHI
water level data in from Monitor. A future card will add any
additional design treatment.

VizLayers is no longer used in the Observation tab, so a note is made
to clean this up in a subsiquent card.

Refs #3243