Only fetch weather stations once
The layer picker updates on model changes, as of e93743eba77.
When we updated the model during layer fetch, it would re-
render the layer picker, which would re-trigger a layer fetch,
which would loop.

Now we only fetch layers if there is an error or there isn't
an active fetch. Furthermore, the "polling" update is done
silently, so it doesn't trigger an extra render.

This follows the doSomethingIfNeeded pattern used elsewhere
in the app.