Highlight Weather Stations used in MapShed Model
This commit adds functionality for highlighting weather stations in the
weather stations layer that are participating in a MapShed modeling
project. This includes the ability to set weather station layer
markers as highlighted and add their distance to the project AOI into
the popup. This functionality is used when the 'gis_data' is fetched
from the MapShed endpoint, which now includes the weather station IDs
used along with the distances. Also added in this PR is a toggle
button on the 'runoff' and 'quality' tabs of the MapShed model result
viewer that allows a user to switch on and off the weather
stations. This differs in the LayerPicker functionality only in that
when toggling on the weather station layer, the map will zoom to a
viewport that allows the user to see the active weather stations and
the AOI.

Refs #3245