Allow more precision on Analyze Coverage tables
With the addition of Protected Lands, the Coverage column in the
Analyze tables could slice off required precision for the display
numbers when the coverage % was less than 1. This more closely matches
the behavior of the horizontal bar charts associated with these tables.

One drawback is that some protected areas are so small that the
rounding used in the sum displayed in the total row may not always
match the sum of the values in the table column, as they be be even
below the 2-digit of precision threshold.