Use preferred unit scheme for Weather Stations
Previously, the mean precipitation was inaccurately reported
in inches, when the value was in millimeters. Now, the value
is standardized and displayed in the user's preferred scheme.

Notably, the popups are rendered when the layer is added. So,
if the user logs in after that, or changes their preferences,
the new scheme isn't reflected in the popups. Resolving this
may involve addressing some long standing issues with design,
which currently has App initialization dependent on the User
model, so the User model cannot trigger items on App.

These issues may be resolved in #3264 and then #3265.