Update Mapbox attribution
According to instructions here:
https://docs.mapbox.com/help/how-mapbox-works/attribution/

Though we're following instructions for Leaflet on that
page, Mapbox may still insist on adding their wordmark to
the map in the bottom-left as well. That space is already
occupied by the Layer Picker, a central and established
widget familiar to all users, so we may have to work with
them on how best to position the wordmark if deemed
mandatory.