Add sample CSV for custom weather data
This CSV will be made available for download when a user
requests it in the front-end. It has three columns:
DATE, PRCP, and TAVG, for the date, precipitation, and
average temperature respectively. The date is in MM/DD/YYYY
format, the precipitation is in inches, and the temperature
in Fahrenheit. These are the units requested by Barry.

However, in practice, we may allow users to specify the
units as they upload the file, or default to the units
set in their preferences.