Add custom weather data fields to Project model
For each GWLF-E project, the user can now possibly upload
Custom Weather Data. This will likely be uploaded to S3 (see #3274).
The Project model is updated to track the key to the upload.
There is also a field to set whether or not to use the custom
weather data, in case the user wants to switch back to the default
weather data without losing that which they've already uploaded.

Also adds migrations.