Generalize ITSI front-end for SSO
In preparation for adding a second Single Sign On provider, the
forms and routes in the front-end are updated to take parameters
for specific providers.

Also rename itsi_enabled to sso_enabled. This is mainly used to
hide ITSI from BiG-CZ portal. By overloading that to be a
universal flag for SSO, we also exclude BiG-CZ from being able
to use Concord logins. However, that is not a likely use case.