Pin llvmlite to 0.31.0 to fix builds
With the release of llvmlite v0.32.0 being auto-picked up by
numba, builds were breaking. By pinning llvmlite to 0.31.0,
the latest working version, we protect ourselves from the
breakage.

We've been using the same version of numba for a few years
now, which started with llvmlite 0.23.0, so we're okay by
not always using the latest version.