Setup Weather dialog template with two columns
One for Available Data, which has a dropdown allowing
selecting between the MapShed in-built source and the
future simulations.

And the second for Custom Data, which has controls for
uploading and getting a sample CSV. This side will also
have controls for exporting and deleting once the file
has been uploaded.