Switch GET to return custom weather JSON
Previously GETing the custom weather endpoint would return
the raw CSV file. However, in cases when the user is switching
from DEFAULT weather type back to CUSTOM, we again need to
parse the original CSV file. For this, we switch the GET of
the custom weather endpoint to return a re-parsed version of
the CSV file.

A new endpoint, custom-weather-data/download is added, for
retrieving the CSV file.

Tests are amended accordingly.