Allow for switching back to CUSTOM from DEFAULT
The weather_type sets the weather_data modification, which is
used to override the gis_data dictionary. In case of DEFAULT,
it is an empty dictionary. In case of CUSTOM, it has fields
for WxYrBeg, WxYrEnd, WxYrs, Precs, and Temps.

When a user switches from CUSTOM to DEFAULT, the weather_data
modification is wiped out. To reinitialize it, we fetch the
custom weather output from the server if the scenario has a
file defined for it. This allows the dialog to have custom
weather data even if it is not present in the modification.