Consolidate Upload Button interaction
Previously upload button's state was set in two different places:
turned off during a fetch, and turned back on during error
handling. This was because we assumed that in the success case
the entire view would be replaced with Existing Weather Data View.

However, in case of re-uploading after validation errors, the
button needs to be turned back on. Furthermore, it is better to
consolidate all the logic concerning it in one place makes it
more composable.