Add, enhance validations for custom weather uploads
The `err` helper method is used as shorthand for reporting
errors. The first parameter is the message, the second is the
line number. If specified, the line number will be prefixed to
the error message.

The validations check for the following:

  - Valid CSV header
  - Valid length of file
  - Valid year range between first and last date
  - All dates are properly parseable
  - There are no duplicate dates
  - There are no missing / out of order dates
  - Precipitation and temperature are valid floats
  - Precipitation is not negative

If a file has too many errors, we only report that first 10.
This is configurable in the MAX_ERRORS variable.