Improve sample CSV, use for testing
The sample CSV is bolstered to have 5 years of values, and be
something that would pass our validations. It's provenance is
`phillyRCP4520802100daily.csv` from Scott Ensign's Future
Weather dataset, with 2080 replaced with 2000.

Furtheremore, we now reuse it for our testing. This reduces
repetition of CSV files in the repository, and ensures that
the sample CSV is valid.

It is a little unusual to use static files for testing, but
they are placed in the right place in the VM during provisioning,
so it works.