Move scenario duplication server side
Previously scenario duplication was done in the client.
Since now we have a field that cannot be directly set
by the client (weather_custom, a file field, for which
the client only ever has a string value), we move the
duplcation logic to the server, with a new $ID/duplicate
endpoint.

While this does duplicate some naming logic from the
client side, it felt cleaner to return a properly
formed duplicate from the server, than have the client
do follow-up tweaks to it and save again.

The client-side `makeNewScenarioName` method is also
used when creating new scenarios, so is left in place.