Save custom weather on every valid change
Previously we would only update weather whent the user clicked
Save. However, we were interacting with, and saving values on,
the server before that. This led to a few scenarios where the
user clicked Cancel, but their changes had already been saved
and were not undone.

By saving on every valid change, and removing Save / Cancel in
favor of a Done button, we better convey the interaction to
the user.