Fix failures in making external storage volumes accessible to RWD
This upgrade addresses an issue with the prior version of azavea.docker
that led to different versions of the Docker daemon and Docker CLI to be
installed.

In addition, this change set removes an override to explicitly use the
AUFS storage driver in development. Instead, we let the Docker installer
chose the best default for the host operating system.