Add back-end for fetching Simulated Weather
If the project is in the DRB, we identify the two nearest
weather stations, and then average their simulated values.
The simulated weather files are in the same format as the
custom weather, so we run them through the same machinery.
Any errors are reported as 500s, since they are the fault
of server side issues. The output is sent with headers to
cache the results, as they won't change.