Allow simulations in DRB only
Make in_drb a property on the project, and pipe it to the
front-end via the serializer. In the front-end, we use this
property to disable the simulation options.

It is possible in the future to have simulations that are not
limited to the DRB. That will require refactoring of this
behavior.

It is possible in the future we need to know if the AoI is in
the DRB before it has been saved as a project. In that case
we'll need to add a similar calculation for `in_drb` in the
front-end. Currently, the Weather Dialog is only available when
the project is saved to the back-end, and will always have a
value for `in_drb` provided from the back-end.