Add Shippensburg DRB 2100 Land Back-end
The analyze API is identical to the rest, and the actual task
is very similar to NLCD's. Just instead of GeoTrellis, the
values are coming from Drexel's Fast Zonal API.

This endpoint is only available in the DRB, and requires an
additional dataset parameter. These requirements are enforced
in the endpoint, making it the only Analyze endpoint with
such data validation. Bad requests get 400s.