Add Shippensburg DRB 2100 Land Front-end
Adds two new analysis types within the Land tab, one for each
kind of DRB 2100 Land available. These are rendered in the same
style as NLCD.

Notably, the layer they are connected to is colored on a
different ramp (white to red based on % urbanization) then the
analysis values, which is an NLCD distribution.