Standardize DRB check
Previously we used two different techniques for checking if a
shape is within the DRB: AOI.within(DRB.buffer(0.1)) in the
back-end, and AOI.intersects(DRB_SIMPLE) in the front-end.

DRB_SIMPLE is a better approximate representation of the shape:
https://user-images.githubusercontent.com/1430060/85214727-15e61a80-b33d-11ea-8617-1d733becf58c.png

We standardize to always doing AOI.within(DRB_SIMPLE) in both
the front- and back-ends.

"Upper Mahanoy Creek" HUC-12 is a good test for the DRB check,
as it only partially overlaps the DRB_SIMPLE_PERIMETER, and
should not qualify for being within the DRB.