Don't re-run analyses in end states
Previously, an analysis would always be re-run if it did not have
a result. This included failed analyses. While this could be
potentially useful, since there is no explicit retry mechanism
for the analyses, supporting one surreptitiously was not the
best designed approach. This caused unnecessary noise when
switching scenarios in the Model phase, as the sidebar is re-
created for every scenario switch, causing failed analyses to
be retried every time.

Until a future time when an explicit retry mechanism is added,
refetching analyses in end states is disabled. An end state is
either one of "failed" or "complete" with a result.