Increase burst throttle rate to 30/min
Since now in the DRB we have 11 analyses, refreshing the page
can resubmit all of those 11, which leads to an amount more
than 20 being submitted in a minute, leading to some requests
being throttled.

Since these throttled requests are not in a end state, they
will be retried by fetchAnalysisIfNeeded. However, to lessen
the chance of throttling, the limit is increased by 50%.