Check for DRB when creating new project
Previously the DRB check was only done on server side,
making it difficult to judge if a project was within
the DRB until after it was saved, and in some cases
even then the correct values did not propogate correctly.

Now we add an evaluation when creating the project that
will ensure that the value is always populated correclty.

This can be done with greater confidence thanks to
0a9e8b9 which standardized the DRB check between the
server and the client.

Refs #3336