WIP
The next step is to add a SPECIAL modification when this preset
changes. That SPECIAL modification should be shifted upon add
to ALWAYS be the FIRST one in the list.

Then, the Land Cover Modal should be updated to MERGE the
SPECIAL modification on load, and on preset change.

Later on, this MERGING of the SPECIAL modification can be
generalized, so that when other modifications such as
Conservation Practices are being added, they resepct the
SPECIAL modification as well.

Furthermore, if modifications are applied in sequence, then
the Land Cover modifications can be respected by others like
the Conservation Practices as well.