Add Landcover Presets
Adds a dropdown at the top of the land cover dialog which
lists available presets. Selecting one of the non-default
presets fetches the value from the related analysis, converts
it into a Mapshed format, and populates the text boxes. Any
existing text box values will be overwritten. The total is
updated accordingly.

In some cases, the total doesn't quite completely match the
value of NLCD. In these cases, the user must adjust the
numbers manually.

On updating the numbers, they are saved as if the user had
input them manually, replicating the existing behavior.