Refactor: Pass in Scenario to Land Cover Dialog
Previously we extracted relevant bits from the scenario and
passed in only those to the Land Cover Dialog. Now, as it
may need to remove modifications, instead of passing in further
bits, we now pass in the entire scenario. This allows for more
flexibility within the Land Cover Dialgo.

Also refactor out App import, which tends to tie things up.