Fix Compare View Crash
If a display name was not available for a certain GWLFE
modification, of which there are many now, including the
land cover preset, and the Settings modifications, the
Compare View would crash.

Now we fallback to the internal key when a display value
is not found. Less than ideal, but does allow the app to
work.