Proportion land cover distribution to match
The raw land cover numbers from the Drexel API don't match
our internal NLCD 2011 numbers, and were slightly off,
resulting in the user having to manually reconcile in the
front-end.

Since these are future projections, the individual raw
numbers are not the main feature, rather the distribution
as a whole. By attuning it to the existing total, we
ensure that the new distribution still reflects the
forecasts but adds up to the same total, resulting in
a more fluid experience.