Add, configure django-cookies-samesite
Recent versions of Chrome require cookies be set with explicit
Cross Site policies. Since MMW is embedded within the LARA
educational system, logging in requires cookies to be tagged
with "SameSite=None; Secure".

This configuration is available out of the box in Django 2.1+,
in order to use it in Django 1.11 we add the django-cookies-samesite
package.

This configuration can cause issues with some older browsers,
a full list of which can be seen here:

https://www.chromium.org/updates/same-site/incompatible-clients

However, since our official policy is to support the two most
recent versions of popular browsers, all of which implement this
functionality correctly, we shouldn't have to worry about them.