Deny empty AoIs
In some cases, projects are mysteriously updated to have NULL
Areas of Interest. Our current API allows this, becuase if a
project has `is_activity=True` then it can, in fact, have empty
Areas of Interest. However, for non-activities, this should not
be the case.

Adding this validation ensures that all non-activity projects
must have a non-NULL area_of_interest specified. The structural
validity of the area_of_interest is inspected by the serializer
automatically, since it is marked as the geo_field.