Allow for project creation with WKAoIs
The front-end currently always creates new projects with an
explicitly specified AoI. However, in most of our other endpoints,
we accept either AoI or WKAoI. The validation is updated to
use the same code which allows indirect specification of AoI
via a valid WKAoI.

There is no plan to update the front-end to make use of this
capability, and this would add a subtle overhead to project
creation. However, in the interest of consistency, this feels
appropriate.