Add database checks preventing empty AoIs
To ensure that AoIs are not deleted in any manner, not just
via the Django Endpoints, we add a database constraint that
checks if a project is an activity, and if it is not ensures
that the Area of Interest is NOT NULL.

This can only be enforced if there are no existing projects
that are non-activities and have NULL AoIs. For that, we
delete all existing projects matching that description.

Before deploying this to production, we will take a manual
back-up of all those projects, in case we need them for
future reference.