Upgrade Vagrantfile for Ruby 3 / Vagrant 2.2.17
Vagrant 2.2.17 adds extended support for Ruby 3, and requires
this new syntax for expanding dictionaries. This new syntax
works with older versions of Vagrant as well as the new one,
so should not require updates on other developer machines or
on Jenkins.

For details, see here: https://github.com/hashicorp/vagrant/issues/12448#issuecomment-876661600