Allow using tokenized layers for geoprocessing
Previously the layer names used for geoprocessing were hard-coded.
As of #3399 we now have a number of NLCD layers, which may be used
for modeling purposes. To replace the NLCD layer in all operations
we switch to this tokenized scheme.

Now the operation definitions only carry the tokens, with a base
configuration defined with the previously used values. Both `run`
and `multi`, the two geoprocessing functions, now take an optional
`layer_overrides` dictionary, which is a mapping from token to
layer name. At run time, these overrides are combined with the
base configuration, which is in turn used to replace all tokens
with real layer names in the operation.