Deepcopy default configuration before tokenizing
A hitherto unnoticed bug was revealed when the Analyze Land
endpoint switched to using tokenized instances of the same
JSON config: the token replacements would change the default
config itself, since dict.copy() is a shallow copy. This is
fixed by using copy.deepcopy() instead.

It is unlikely that this would have occurred in the past, as

  1. We used a separate JSON block for every geoprocessing
     run, rather than the same blocks as we are now with
     tokenization
  2. The contents of the JSON block were not changed other
     than adding the input polygon, which would be overwritten
     on every call