Add layer_overrides JSON field to Project model
This field will be used to save the layer overrides for each
project. All existing projects are initialized with NLCD 2011,
and all new projects will be initialized with an empty override
dictionary.

We use a JSONField which is available as of Django 1.11. This
is a better representation than TextField which is what was used
so far.