Mark land covers as loading if still analyzing
Since we use the Analyze results for each land cover type to
populate the values in this dialog, we need to wait for that
to finish before those values can be used. In case some of
the analyses are still in progress, those options are now
disabled, with helpful messages indicating why that is the
case.