Install Python 2.7 on Ubuntu 20.04 Focal
Ubuntu 20.04 Focal does not have a `python` package in its apt
repository. Instead, it has a `python2` package. Since this is
an unusual case, of installing an older version of Python in a
newer OS, we switch from using the generic `azavea.python` and
`azavea.pip` roles to an MMW specific one, which installs the
most recent version of Python 2.7, and symlinks `python2` to
`python`, making the default version by Python 2.7.

The most recent version of pip that supports Python 2.7 is also
installed.