Hack Python 2.7 / Django 1.11 for new GEOS
This version of GEOS has a mistake in its version number which
ends with an empty space. Newer versions of libgeos can handle
that, but the one we're using with Django 1.11 and Python 2.7
does not. There is a bug for this: https://code.djangoproject.com/ticket/31838,
but the advice there is to upgrade Python and Django.

We are not ready to do that just yet, so instead we fix
the version checking line in the library with an additional
strip() which removes any extraneous spaces.