Generalize streams collection for geoprocessing
The values in the medium resolution table nhdflowline are LineStrings,
which when passed through ST_Collect result in a MultiLineString. The
new table nhdflowlinehr has rows of type MultiLineString, which when
passed through ST_Collect result in a GeometryCollection.

The Geoprocessing Service API expects an array of stringified GeoJSON
of MultiLineStrings. Previously we would send a singleton array with
a MultiLineString which was the result of ST_Collect.

Now we send a list of strings, where each string is a MultiLineString,
which we get via ST_Multi. This converts LineStrings to MultiLineStrings,
but keeps MultiLineStrings as is, ensuring that the final array is the
same regardless of datasource, and works with the existing API.