Specify datasource in analyze streams API path
This is a breaking change, as this API now requires an explicit
specification of the stream datasource. A default was not chosen
for similar reasons as for the NLCD case in bb107ae2, namely that
the current default is not going to be the default going forward.

Currently, the NHD Medium Resolution datasource was the default,
the equivalent of analyze/streams/nhd. But going forward, the NHD
High Resolution datasource is going to be the default, the equivalent
of analyze/streams/nhdhr. Because of this changing default, we don't
specify any default value, instead asking the user to be explicit
about the ones they want.

The datasource is also added to the geoprocessing cache_key, so that
different layers can be cached independently.