Use clipped stream lengths in analysis
Previously we summed the full length of any stream that
intersected with the area of interest. This led to very
large lengths being reported even for small areas of interest,
and unexpected values when comparing medium resolution to
high resolution streams.

Now we fix this calculation by first clipping the streams
to the area of interest, and then summing their lengths.

Refs #3430