Use arrays for stream lines in /multi jobs
Previously we used an array for stream lines in /run jobs,
and a single MultiLineString in /multi jobs. This was fine
as long as we only ever sent one MultiLineString. However,
as of 4f485777 we send an array of MultiLineStrings. Thus,
the /multi endpoint is updated to support them as well.

Also see: https://github.com/WikiWatershed/mmw-geoprocessing/pull/94