Increase geoprocessing timeout for HUC-8 subbasin
Certain HUC-8s were timing out for subbasin analysis because
of the new hi res streams which take a longer time to process.

This pulls in a new version of the geoprocessing service, which
includes a larger timeout limit of 120s. This is also applied
to the Celery Task Timeout, so the tasks wait for the long
requests.

Incorporates work done in https://github.com/WikiWatershed/mmw-geoprocessing/pull/98