Add vagrant-disksize, increase services to 64GB
The hi res streams data is very large, and often fails import into
the default 30GB disk. This plugin allows for setting the disk size
in the Vagrant file, and we add a shell provisioner to increase the
logical volume and file system to use it.

The README is updated accordingly.