Set geoprocessing timeout as env var, up to 200s
With the latest version of the geoprocessing service, we can now
override certain settings with environment variables. We had set
the timeout to 160s in the last version, but that is not enough,
so we up it to 200s. In the future, we can set this simply by
setting this value, without needing a full re-deploy of the
geoprocessing service.