Optimize Stream Intersection Calculation
Previously we would calculate the intersection of all streams
within an area of interest and sum their lengths. For large
areas of interest, especially with the high resolution streams,
this can be an expensive operation because of the number of
streams they cover.

To optimize this, for each intersecting stream, we check if it
is entirely covered by the area of interest, using ST_CoveredBy.
This is a much faster operation than calculating the intersection
via ST_Intersection. If it is completely covered, we use the full
stream. If it is not completely covered, then we calculat the
intersection.

The speedup offered by this depends on the size of the AoI and
the number of streams within it. As a benchmark I used the
Schuylkill HUC-8. Before this change the stream calculation
query took about 40s, now it takes less than 20s.

For more details, see: https://postgis.net/2014/03/14/tip_intersection_faster/