Add method to get WKAoI given a HUC
Given a valid HUC, composes the appropriate WKAoI. The HUC
is checked for validity, to ensure it is the right length,
and to ensure it matches a row in the database.

We translate a HUC into a WKAoI because of the existing
infrastructure (caching, code) to support WKAoIs. HUCs are
more widely known in the industry at large, and will allow
more third parties to take advantage of our caching.