Don't use iframes for HydroShare login
Previously we used iframes for logging users into HydroShare,
which was convenient and friendly, and worked in all modern
browsers. For IE which did not allow third party cookies in
iframes, necessary for HydroShare login, we fellback to a
full-page experience.

As of recent times, Chrome no longer allows third party
cookies in iframes either. So, we remove the iframe flow
entirely, and replace all of it with the fallback. Not as
elegant as before, but functional.

Refs #3440