Handle input for subbasin prepare endpoint
Subbasin is only allowed on HUC-8s, HUC-10s, and HUC-12s.
Thus, this endpoint does not take arbitrary GeoJSON as input.
The selected shape may be provided as WKAoI or HUC, with
optional layer_overrides.

If neither WKAoI nor HUC is provided, a 400 is raised.
Also if WKAoI or HUC are invalid, or if no HUC-12s are found
for a valid WKAoI (such as Mimbres HUC-8 huc8__253).