Add job_uuid to job submitted response
Everywhere else we use `job_uuid`, but in this case we used
`job`. This was odd. Now `job_uuid` is added to the job
submitted response as well, ensuring it is used uniformly
throughout. The documentation is updated to only list the
new `job_uuid` field.

The old `job` field is still present in the output, will be
removed in a later release. A deprecation notice is added so
that users have time to switch from `job` to `job_uuid`.