Add ability to override taskName via helper
We've used the taskHelper to override many of the taskModel's
attributes. Now, we come across a situation where the taskName
needs to be overridden. Previously we used the same endpoint to
retrieve Mapshed and Subbasin results, with the difference
marked via query parameter. Now we use separate endpoints. Thus,
the taskHelper needs to override the taskModel's URL. Hence
this overridability is added.