Add Validation to gwlf-e/run Endpoint
Make sure `input` and `job_uuid` are not specified together, and that
`input` has all the required valid keys to run the modeling.

We are not validating the value of the input in /run/ since there are
more than 400 keys in the input object.

A follow-up issue will be added to call out specific missing keys from
the payload when checking the `input` of `gwlf-e/run/`