Validate key presence before validating values
Previously the validation of gwlf-e/run input keys and values
was combined in one function. This made it so that if only the
job_uuid was provided, and the model_input retrieved from that,
it would mark the input as invalid.

Now, we first check for the keys separately. If they are not
in the right manner, we stop quickly. At the end, when the values
are validated, it does not matter if they came from the user or
the database via job_uuid. They will be validated equally.