Validate and indicate missing keys from input
The full list of expected gwlf-e/run input was made by
taking a correct output, removing the keys from GWLFE_DEFAULT
to identify the remaining ones, dropping watershed_id which is
only there for subbasin runs, and using the rest to make a
full list of expected keys for gwlf-e/run.

Any keys missing from the input are called out.

This works both for user provide and MMW-generated input.