Set TR-55 to use NLCD11 2011 by default
As noted in #3527, the NLCD19 + Soil combination for
RasterGroupedCount would result in unstable values,
reporting NODATA for cases that had real values.

While investigating this failure, it was observed
that the NLCD11 2011 + Soil combination worked
correctly, as before. After conferring with clients,
it was decided that TR-55 should be switched back
to NLCD11 2011, until further notice.

This code change allows the specification of a land
layer override, but in the absence of an override
will use the NLCD11 2011 layer. Since the UI does
not specify any overrides, the NLCD11 2011 layer
will be used for all cases.