Add Catchment-level Source Loads Fields in Subbasin/Run
We got new fields from the new SRAT API and we included the new fields
to subbasin/run according to the investigation in #3521.

A new 'Loads' key is added to include the detailed T/P/Sediment loads
from different sources at catchment level.

curlify is added to make endpoint testing easier in the future.