Add `in_drwi` attribute to Project Model
This is not the most sustainable design, but will do for now.
Just like the `in_drb` attribute, we add a new one that checks
if the project is in the DRWI region.

This is also added to the weather modal's WindowModel, since
that will need this information for correct rendering of
simulated weather options.