Add new draw API endpoint for drainage areas
Since drainage areas will need API interaction to draw correctly,
we add a new family of APIs nested under the draw/ path. This
dummy implementation returns a buffered shape around the given
point, but will eventually be replaced with an API call to the
Drexel / ANS endpoint.

The front-end is updated to call this API when using the point-
based drainage area draw tool. This is currently a naive POST,
but may be converted to a taskRunner based implementation, a-la
RWD, if the Drexel / ANS endpoint ends up taking a while to
respond.