Add simple stream-based drainage area draw tool
Similar to the point one, this also sends a request to the
API, and shows the returned shape.

When this draw tool is activated, the High Resolution Streams
layer is turned on, so that the user can draw an area around
a stream.

This implementation currently does not have any error handling,
or support for delays. It is likely that we'll need to transition
these endpoints to a Celery task based workflow, as is the case
for all the other Geoprocessing API endpoints.