Refactor: Make drainage area endpoints task-based
Previously the endpoints were made as simple request-response
functions. However, given that calls to the third party API
may take a while, it is better to make them task-based, as all
other public APIs are, and return the results via the job
endpoint.

The front-end is updated correspondingly to show a spinner and
error messages as appropriate while waiting for the task to
finish.